Adding Animations to player's class
some weird exceptions apears

+++ Source/Physic/Rigidbody/Rigidbody.hpp
@@ -20,9 +20,9 @@ private:
 	size_t m_physicEngineID;
 
 public:
-	int positionX;
-	int positionY;
-	int velocityX;
-	int velocityY;
+	float positionX;
+	float positionY;
+	float velocityX;
+	float velocityY;
 };
